SCHED-125: Fix rebase issues and add changes to adapt new structure
