+++ components/optimizer/__init__.py
@@ -1,5 +1,5 @@
 from components.collector import NightEvents
-from components.optimizer.base import Selection
+from common.calculations.selection import Selection
 from .greedymax import GreedyMax
 from common.minimodel import Plans
 
@@ -11,17 +11,15 @@ class Optimizer:
     """
 
     def __init__(self, selection: Selection, algorithm=GreedyMax(some_parameter=1)):
-        self.algorithm = algorithm.add(selection)
+        self.algorithm = algorithm.add(selection.program_info)
+        self.night_events = selection.night_events
     
-    def schedule(self, night_events: NightEvents) -> Plans:
+    def schedule(self) -> Plans:
         # TODO: This forces to make Plans for each site SEPARATELY.
         # This would create and issue with OR groups as observations can be schedule
         # on different sites but not twice. Old GM had some check-ins to handle this
         # cases but right now we don't have OR groups implemented.
 
-        plans = Plans(night_events)
+        plans = Plans(self.night_events)
         self.algorithm.schedule(plans)
         return plans
-        
-        
-
