+++ components/optimizer/dummy.py
@@ -15,23 +15,23 @@ class DummyOptimizer(BaseOptimizer):
         Gives a random group/observation to add to plan
         """
         while all(plan.is_full() for plan in plans):
-
-            ran_group = random.choice(list(self.selection.values()))
-            for observation in ran_group[0].observations():
-                # TODO: This should be constant not lineal time
-                for plan in plans:
-                    if not plan.is_full and plan.site == observation.site:
-                        obs_len = plan.time2slots(observation.total_used())
-                        if (plan.time_left() >= obs_len):
-                            plan.add(observation, obs_len)
-                            break
-                        else:
-                            # TODO: DO a partial insert
-                            # Splitting groups is not yet implemented
-                            # Right now we are just going to finish the plan
-                            plan.is_full = True
+            for program in self.programs:
+                ran_group = random.choice(program.group_data.values())
+                for observation in ran_group.group.observations():
+                    # TODO: This should be constant not lineal time
+                    for plan in plans:
+                        if not plan.is_full and plan.site == observation.site:
+                            obs_len = plan.time2slots(observation.total_used())
+                            if (plan.time_left() >= obs_len):
+                                plan.add(observation, obs_len)
+                                break
+                            else:
+                                # TODO: DO a partial insert
+                                # Splitting groups is not yet implemented
+                                # Right now we are just going to finish the plan
+                                plan.is_full = True
         
-    def add(self, selection: Selection):
+    def add(self, programInfo: Selection):
         # Preparation for the optimizer i.e create chromosomes, etc.
-        self.selection = selection
+        self.programs = [p for p in programInfo.values()]
         return self
